requirements.txt
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-tenant-schemas==1.10.0
psycopg2-binary==2.9.7
redis==5.0.1
celery==5.3.4
Pillow==10.1.0
python-decouple==3.8
django-filter==23.3
drf-spectacular==0.26.5
django-debug-toolbar==4.2.0
gunicorn==21.2.0
whitenoise==6.6.0
django-storages==1.14.2
boto3==1.34.0
requests==2.31.0
python-dotenv==1.0.0 